backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
passlib[bcrypt]
chromadb
google-generativeai
langchain
sentence-transformers
fitz